@@ -3,20 +3,24 @@
   "version": "1.0.0",
   "description": "tell storys online",
   "main": "app.js",
-  "engines": { 
-    "node": "14.x"
+  "engines": {
+    "node": "16.x"
   },
   "scripts": {
-    "start": "cross-env NODE_ENV=production node app",
-    "dev": "cross-env NODE_ENV=development nodemon app"
+    "start": "cross-env NODE_ENV=production node /build/app.js",
+    "dev": "concurrently \"npx tsc --watch\" \"nodemon ./src/app.ts\"",
+    "copy-static-assets": "node copyStaticAssets.js",
+    "build-ts": "tsc",
+    "build": "npm run build-ts && npm run copy-static-assets"
   },
   "author": "Armin Srabovic",
   "license": "MIT",
   "dependencies": {
     "connect-mongo": "^4.6.0",
     "dotenv": "^16.0.0",
+    "ejs": "^3.1.8",
     "express": "^4.17.3",
-    "express-handlebars": "^6.0.3",
+    "express-handlebars": "^6.0.7",
     "express-session": "^1.17.2",
     "html-to-text": "^8.2.0",
     "method-override": "^3.0.0",
@@ -26,7 +30,20 @@
     "passport-google-oauth20": "^2.0.0"
   },
   "devDependencies": {
+    "@types/express": "^4.17.17",
+    "@types/express-session": "^1.17.6",
+    "@types/html-to-text": "^9.0.0",
+    "@types/method-override": "^0.0.32",
+    "@types/mongoose": "^5.11.97",
+    "@types/morgan": "^1.9.4",
+    "@types/node": "^18.14.1",
+    "@types/passport": "^1.0.12",
+    "@types/passport-google-oauth20": "^2.0.11",
+    "@types/shelljs": "^0.8.11",
+    "concurrently": "^7.6.0",
     "cross-env": "^7.0.3",
-    "nodemon": "^2.0.15"
+    "nodemon": "^2.0.20",
+    "shelljs": "^0.8.5",
+    "ts-node": "^10.9.1"
   }
 }
