@@ -1,21 +1,20 @@
-const express = require("express");
-const { redirect } = require("express/lib/response");
-const router = express.Router();
-const { ensureAuth } = require("../middleware/auth");
-const Story = require("../models/Story");
+import { RequestHandler } from "express";
+import Story from "../models/Story";
+import { IUser } from "../models/User";
+
 /**
  * @desc Show Add page
  * @route GET /stories/add
  */
-router.get("/add", ensureAuth, (req, res) => {
+export const add: RequestHandler = async (req, res, next) => {
   res.render("../views/stories/add.hbs");
-});
+};
 
 /**
  * @desc Show Stories page
  * @route GET /stories
  */
-router.get("/", ensureAuth, async (req, res) => {
+export const stories: RequestHandler = async (req, res, next) => {
   try {
     const stories = await Story.find({ status: "public" })
       .populate("user")
@@ -27,13 +26,13 @@ router.get("/", ensureAuth, async (req, res) => {
     console.log(error);
     res.render("error/500");
   }
-});
+};
 
 /**
  * @desc Show single Story
  * @route GET /stories/:id
  */
-router.get("/:id", ensureAuth, async (req, res) => {
+export const story: RequestHandler = async (req, res, next) => {
   try {
     let story = await Story.findById({ _id: req.params.id })
       .populate("user")
@@ -44,52 +43,58 @@ router.get("/:id", ensureAuth, async (req, res) => {
     console.log(error);
     res.render("error/500");
   }
-});
+};
 
 /**
  * @desc Process added Story
  * @route POST /stories/
  */
-router.post("/", ensureAuth, async (req, res) => {
+export const addStory: RequestHandler = async (req, res, next) => {
   try {
-    req.body.user = req.user.id;
+    const user = req.user as IUser;
+    req.body.user = user.id;
+
     await Story.create(req.body);
     res.redirect("/dashboard");
   } catch (error) {
     console.log(error);
     res.render("../views/errors/500");
   }
-});
+};
 /**
  * @desc Edit a Story
  * @route GET /stories/edit/:id
  */
-router.get("/edit/:id", ensureAuth, async (req, res) => {
+export const editStory: RequestHandler = async (req, res, next) => {
   try {
     const story = await Story.findOne({ _id: req.params.id }).lean();
 
     if (!story) return res.render("error/404");
 
-    story.user != req.user.id
+    const user = req.user as IUser;
+
+    story.user != user.id
       ? res.redirect("/stories")
       : res.render("stories/edit", { story });
   } catch (error) {
     console.log(error);
     res.redirect("error/500");
   }
-});
+};
 
 /**
  * @desc Update a Story
  * @route PUT /stories/:id
  */
-router.put("/:id", ensureAuth, async (req, res) => {
+export const updateStory: RequestHandler = async (req, res, next) => {
   try {
     let story = await Story.findById(req.params.id).lean();
 
     if (!story) return res.render("error/404");
 
-    story.user != req.user.id
+    const user = req.user as IUser;
+
+    story.user != user.id
       ? res.redirect("/stories")
       : (story = await Story.findOneAndUpdate(
           { _id: req.params.id },
@@ -105,13 +110,13 @@ router.put("/:id", ensureAuth, async (req, res) => {
     console.log(error);
     res.redirect("error/500");
   }
-});
+};
 
 /**
  * @desc Delete Story
  * @route DELETE /stories/:id
  */
-router.delete("/:id", ensureAuth, async (req, res) => {
+export const deleteStory: RequestHandler = async (req, res, next) => {
   try {
     await Story.remove({ _id: req.params.id });
     res.redirect("/dashboard");
@@ -119,13 +124,13 @@ router.delete("/:id", ensureAuth, async (req, res) => {
     console.log(error);
     res.redirect("error/500");
   }
-});
+};
 
 /**
  * @desc User Stories
  * @route GET /stories/user/:userId
  */
-router.get("/user/:userId", ensureAuth, async (req, res) => {
+export const userStories: RequestHandler = async (req, res, next) => {
   try {
     let stories = await Story.find({
       user: req.params.userId,
@@ -138,5 +143,4 @@ router.get("/user/:userId", ensureAuth, async (req, res) => {
     console.log(error);
     res.redirect("error/500");
   }
-});
-module.exports = router;
+};
